misc/requirements.txt

hidapi
adafruit-blinka
